Added link to the spec repository
The original link pointed to the (now archived) confluence page. Right now the spec link points to the master branch of the spec repository. Should it point to develop instead for the develop branch of this repository, and have the master branch of this repository point to the tagged release of the spec (once it's stable)?