remove 'file:' prefix from fixture location pattern
This allows us to include the prefix in the config, which is more flexible.
For example, we can set file:fixtures for the default fixtures,
which need to be overridable in the docker container,
and we can set classpath:test-fixtures for the test fixtures,
which can then be included in the test/resources dir.

Moreover, this approach is similar to the way flyway.locations are specified.