update fixture history based on populator resources
This uses reflection to get access to the private resources field.
A bit of a hack, but an alternative based on subclassing would probably be more cumbersome.
We need this to handle the case where bootstrapping is disabled but we still use the resetservice.
Alternatively, we could simply remove the ability to disable bootstrapping, which would simplify things greatly.